@@ -1,19 +1,20 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
 import 'package:fluttericon/font_awesome5_icons.dart';
-
 import 'package:zhalkhas_githubio/text.dart';
 import 'package:zhalkhas_githubio/util.dart';
 
 class Profile extends StatelessWidget {
+  const Profile({Key? key}) : super(key: key);
+
   @override
-  Widget build(BuildContext buildContext) {
+  Widget build(BuildContext context) {
     return Column(
       mainAxisSize: MainAxisSize.max,
       mainAxisAlignment: MainAxisAlignment.spaceAround,
       children: [
         Card(
-          shape: CircleBorder(),
+          shape: const CircleBorder(),
           elevation: 15,
           child: CircleAvatar(
             radius: 95,
@@ -25,7 +26,7 @@ class Profile extends StatelessWidget {
           style: heading,
           textAlign: TextAlign.center,
         ),
-        Text(
+        const Text(
           "Senior year student at Nazarbayev University, Astana"
           "\nJunior Software Engineer",
           textAlign: TextAlign.center,
@@ -35,34 +36,47 @@ class Profile extends StatelessWidget {
           children: [
             Card(
               elevation: 5,
-              child: FlatButton.icon(
-                icon: Icon(FontAwesome5.github
-                    //color: Colors.white,
-                    ),
-                onPressed: redirect('https://github.com/Zhalkhas'),
-                label: Text("Github"),
+              child: TextButton.icon(
+                icon: const Icon(
+                  FontAwesome5.github,
+                  color: Colors.white,
+                ),
+                onPressed:
+                    redirect('https://github.com/Zhalkhas') as void Function()?,
+                label: const Text(
+                  "Github",
+                  style: TextStyle(color: Colors.white),
+                ),
               ),
             ),
             Card(
               elevation: 5,
-              child: FlatButton.icon(
-                icon: Icon(
+              child: TextButton.icon(
+                icon: const Icon(
                   FontAwesome5.telegram_plane,
                   color: Colors.blue,
                 ),
-                onPressed: redirect('https://t.me/kruzenshtern2'),
-                label: Text("Telegram"),
+                onPressed:
+                    redirect('https://t.me/kruzenshtern2') as void Function()?,
+                label: const Text(
+                  "Telegram",
+                  style: TextStyle(color: Colors.white),
+                ),
               ),
             ),
             Card(
               elevation: 5,
-              child: FlatButton.icon(
-                icon: Icon(
+              child: TextButton.icon(
+                icon: const Icon(
                   Icons.mail_outline,
                   color: Colors.red,
                 ),
-                onPressed: redirect('mailto:zhalgas.khassenov@nu.edu.kz'),
-                label: Text("Gmail"),
+                onPressed: redirect('mailto:zhalgas.khassenov@nu.edu.kz')
+                    as void Function()?,
+                label: const Text(
+                  "Gmail",
+                  style: TextStyle(color: Colors.white),
+                ),
               ),
             ),
           ],
