@@ -3,10 +3,12 @@ import 'package:responsive_framework/responsive_framework.dart';
 import 'package:zhalkhas_githubio/screens/base.dart';
 
 void main() {
-  runApp(MyApp());
+  runApp(const MyApp());
 }
 
 class MyApp extends StatelessWidget {
+  const MyApp({Key? key}) : super(key: key);
+
   // This widget is the root of your application.
   @override
   Widget build(BuildContext context) {
@@ -14,18 +16,18 @@ class MyApp extends StatelessWidget {
       title: 'zhalkhas.github.io',
       theme: ThemeData.dark(),
       builder: (context, widget) => ResponsiveWrapper.builder(
-        widget,
+        widget!,
         maxWidth: 1200,
         minWidth: 480,
         defaultScale: true,
         breakpoints: [
-          ResponsiveBreakpoint.autoScale(480, name: MOBILE),
-          ResponsiveBreakpoint.autoScale(800, name: TABLET),
-          ResponsiveBreakpoint.resize(1000, name: DESKTOP),
+          const ResponsiveBreakpoint.autoScale(480, name: MOBILE),
+          const ResponsiveBreakpoint.autoScale(800, name: TABLET),
+          const ResponsiveBreakpoint.resize(1000, name: DESKTOP),
         ],
       ),
       initialRoute: '/',
-      routes: {'/': (context) => Base()},
+      routes: {'/': (context) => const Base()},
     );
   }
 }
